ReactTestRenderer: inline component creation
